perf: separate dashboard DTOs
So that they can be imported without the entire DashboardPresenter
