+++ src/practices/IPractice.ts
@@ -2,8 +2,7 @@ import { PracticeEvaluationResult } from '../model';
 import { PracticeContext } from '../contexts/practice/PracticeContext';
 import { ReportTable, ReportText } from '../reporters/ReporterData';
 import { FixerContext } from '../contexts/fixer/FixerContext';
-import { SecurityIssueDto } from '..';
-import { SecurityIssueSummaryDto, UpdatedDependencyDto, LinterIssueDto, PullRequestDto } from '../reporters';
+import { SecurityIssueSummaryDto, UpdatedDependencyDto, LinterIssueDto, PullRequestDto, SecurityIssueDto } from '../reporters';
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export interface IPractice<T extends {} = {}> {
