+++ src/reporters/DashboardReporterEnums.ts
@@ -1,0 +1,18 @@
+export enum SecurityIssueSeverity {
+  Info = 'info',
+  Low = 'low',
+  Moderate = 'moderate',
+  High = 'high',
+  Critical = 'critical',
+}
+
+export enum UpdatedDependencySeverity {
+  Low = 'low',
+  Moderate = 'moderate',
+  High = 'high',
+}
+
+export enum LinterIssueSeverity {
+  Warning = 'warning',
+  Error = 'error',
+}
